refactor(image-annotation): :recycle: move the lasso code to its own file and removed react native canvas
